+++ Q3E/src/main/jni/doom3/neo/idlib/Lexer.cpp
@@ -3495,7 +3495,7 @@ int Lexer::ReadToken(idToken *token)
 #endif
 
 							//ltoa(token->intvalue, buffer, 10);
-							idStr::snPrintf( buffer, buffersize, "%ld", token->intvalue );
+							idStr::snPrintf( buffer, buffersize, "%d", token->intvalue );
 							assert(token->intvalue == atol(buffer));
 							*token = buffer;
 							token->subtype = TT_INTEGER | TT_DECIMAL | TT_VALUESVALID;
@@ -3531,8 +3531,8 @@ int Lexer::ReadToken(idToken *token)
 #endif
 
 							//ultoa(token->intvalue, buffer, 10);
-							idStr::snPrintf( buffer, buffersize, "%lu", token->intvalue );
-							assert(token->intvalue == ((unsigned long)atol(buffer)));
+							idStr::snPrintf( buffer, buffersize, "%u", token->intvalue );
+							assert(token->intvalue == atol(buffer));
 							*token = buffer;
 							token->subtype = TT_INTEGER | TT_UNSIGNED | TT_DECIMAL | TT_VALUESVALID;
 							break;
