+++ Q3E/src/main/jni/doom3/neo/mod/doom3/sabot/gamesys/Class.cpp
@@ -320,7 +320,7 @@ idClass::FindUninitializedMemory
 void idClass::FindUninitializedMemory(void)
 {
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-    unsigned long *ptr = ((unsigned long *)this) - 1;
+    unsigned int *ptr = ((unsigned int *)this) - 1; // 64long
     int size = *ptr;
     assert((size & 3) == 0);
     size >>= 2;
@@ -512,7 +512,7 @@ void *idClass::operator new(size_t s)
     numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-    unsigned long *ptr = (unsigned long *)p;
+    unsigned int *ptr = (unsigned int *)p; // 64long
     int size = s;
     assert((size & 3) == 0);
     size >>= 3;
@@ -538,7 +538,7 @@ void *idClass::operator new(size_t s, int, int, char *, int)
     numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-    unsigned long *ptr = (unsigned long *)p;
+    unsigned int *ptr = (unsigned int *)p; // 64long
     int size = s;
     assert((size & 3) == 0);
     size >>= 3;
